+++ services/horizon/internal/ingest/main.go
@@ -158,7 +158,7 @@ type System interface {
 	Metrics() Metrics
 	StressTest(numTransactions, changesPerTransaction int) error
 	VerifyRange(fromLedger, toLedger uint32, verifyState bool) error
-	ReingestRange(fromLedger, toLedger uint32, force bool) error
+	ReingestRange(ledgerRanges [][2]uint32, force bool) error
 	BuildGenesisState() error
 	Shutdown()
 }
@@ -486,23 +486,52 @@ func (s *system) VerifyRange(fromLedger, toLedger uint32, verifyState bool) erro
 	})
 }
 
+func validateRanges(ledgerRanges [][2]uint32) error {
+	for i, pair := range ledgerRanges {
+		from, to := pair[0], pair[1]
+		if from > to {
+			return errors.Errorf("Invalid range: %v from > to", pair)
+		}
+		if from == 0 {
+			return errors.Errorf("Invalid range: %v genesis ledger starts at 1", pair)
+		}
+		if i == 0 {
+			continue
+		}
+		prev := ledgerRanges[i-1]
+		if prev[1] >= from {
+			return errors.Errorf("ranges are not sorted prevRange %v  curRange %v", prev, pair)
+		}
+	}
+	return nil
+}
+
 // ReingestRange runs the ingestion pipeline on the range of ledgers ingesting
 // history data only.
-func (s *system) ReingestRange(fromLedger, toLedger uint32, force bool) error {
-	run := func() error {
-		return s.runStateMachine(reingestHistoryRangeState{
-			fromLedger: fromLedger,
-			toLedger:   toLedger,
-			force:      force,
-		})
+func (s *system) ReingestRange(ledgerRanges [][2]uint32, force bool) error {
+	if err := validateRanges(ledgerRanges); err != nil {
+		return err
 	}
-	err := run()
-	for retry := 0; err != nil && retry < s.maxReingestRetries; retry++ {
-		log.Warnf("reingest range [%d, %d] failed (%s), retrying", fromLedger, toLedger, err.Error())
-		time.Sleep(time.Second * time.Duration(s.reingestRetryBackoffSeconds))
-		err = run()
+	for _, pair := range ledgerRanges {
+		fromLedger, toLedger := pair[0], pair[1]
+		run := func() error {
+			return s.runStateMachine(reingestHistoryRangeState{
+				fromLedger: fromLedger,
+				toLedger:   toLedger,
+				force:      force,
+			})
+		}
+		err := run()
+		for retry := 0; err != nil && retry < s.maxReingestRetries; retry++ {
+			log.Warnf("reingest range [%d, %d] failed (%s), retrying", fromLedger, toLedger, err.Error())
+			time.Sleep(time.Second * time.Duration(s.reingestRetryBackoffSeconds))
+			err = run()
+		}
+		if err != nil {
+			return err
+		}
 	}
-	return err
+	return nil
 }
 
 // BuildGenesisState runs the ingestion pipeline on genesis ledger. Transitions